# perf: Optimize async method allocations