This is a rework of a closed PR from corefx, https://github.com/dotnet/corefx/pull/37254 and I wrote a small essay explaining why it was a good idea in the initial post for that so it's worth clicking through and reading it.

Profiling the [DataAccessPerformance](https://github.com/aspnet/DataAccessPerformance) project which emulates the [TechEmpower fortunes benchmark](https://www.techempower.com/benchmarks/#section=data-r17&hw=ph&test=fortune) you can see that some of the top allocations are state closures for async methods like `ReadAsync` `GetFieldValueAsync`. Investigating this I found the code to be much more complex than I'd expected and quite confusing. In order to reduce the allocations I reworked the async infrastructure to use concrete classes and then cached some which are used repeatedly.

Before:
![master](https://user-images.githubusercontent.com/13322696/69499637-0f4d0d00-0eec-11ea-89cb-c26e81e61d80.PNG)

After:
![branch](https://user-images.githubusercontent.com/13322696/69499638-12e09400-0eec-11ea-8f68-cd951aacda7c.PNG)

DataAccessPerformance results:

| name | sync | threads | TPS  | stdev | description |
| ---- | ---- | ------- | ---: | ----: | :---------- |
|ado-sqlclient+async+128|async|128|64667|3262|master|
|ado-sqlclient+async+64|async|64|69262|1592|master|
|ado-sqlclient+async+32|async|32|68814|1918|master|
|ado-sqlclient+async+128|async|128|67288|1595|branch|
|ado-sqlclient+async+64|async|64|68870|2395|branch|
|ado-sqlclient+async+32|async|32|68311|1949|branch|


Which shows a ~5% throughput improvement in situations where CPU time is a limitation and less improvements where it isn't a limitation. I recommend reviewing manually because the diff makes the large number of changes and the extraction of lambdas look particularly confusing.